+++ shopme-common/src/main/java/com/shopme/common/entity/ShippingRate.java
@@ -15,7 +15,7 @@ public class ShippingRate {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Integer id;
 
-	private float rate;
+	private int rate;
 	private int days;
 	
 	@Column(name = "cod_supported")
